ReferenceOrValue class implemented
The ReferenceOrValue class is a good way to create a wrapper for a value
or a reference. It is usefull in contexts like UI, it prevents creating
events and listener to keep, for example, a UI widget up-to-date with a
component value. Instead, we can use some ReferenceOrValue in our UI widgets
and in the editor we can assign a reference to component values to them.