Open with Overload supported
In Windows, you can right click on a file and select "Open With...".
This features add the possibility to directly open a .ovproject with
Overload. The project path (xxx/xxx/xxx.ovproject) is automatically
received as second argument in the main of a C++ project (The first
argument being the .exe path itself). The feature simply consist into
checking the second argument (argv[1]) and, if it exists, skip the
project hub to directly open the project. I had to change the working
directory to always be at the .exe location, otherwise the program won't
find the resources necessary to launch the engine.