Light 3D icons has been added to the scene view
Some texture has been added to the RawTexture.h file and are now
registered has editor resources. These texture are drawn as billboard
(Billboard raw shader has been added to the editor). We also render
lights as non-textured plane for actor picking using the color matching
with the actor ID as the diffuse.