Implementing a simple hierarchy finder
We can now search for an actor by name with the hierarchy search bar.
Actors that matches the input string are shown in the hierarchy and
other actors are hidden. I also noticed m_widgetActorLink was never
cleared so I added it. It seems to be fixing the duplication crash that
we had.