Adding a feedback when hovering an actor or a gizmo
Picking a gizmo axis was sometimes tricky as the collider was small.
The collider has been changed for a bigger and less error-prone one (Arrow_Picking model).

In order to give our users more feedback about their actions, an
highlight outline has been added. This outline will be shown everytime
an actor or a gizmo axis is under the mouse cursor.