Replacing all tabs per spaces
Tabs are poorly interpreted by some applications (GitHub, VIM...). Their
size is dependent of the application and can cause display issues.
In order to make the whole repository more readable (Via GitHub.com), we
should adopt spaces instead of tabs.
This is a breaking change, but must be done early.