Reversing the order of entries in profiler reports
The ProfilerReport is containing a vector of `Action`. These actions
were sorted by duration (From lowest to highest duration). However, when
profiling, the longest action is generally the one we care more about,
thus, actions are now stored from the highest duration to the lowest
one. As a result, the profiler results are now shown in the editor in
the correct order (Longest actions on top, shortest actions on bottom).