Updating editor style
A new editor style is introduced by this commit. The overall design is
flatter, with less curves and more grey tints.

A non-scrollable setting has also been added to `PanelWindow`, which is
used by every views (A view is just texture, render target of a scene
from the point of view of a specific camera). Sometimes when resizing a
view, a scrollbar was temporary shown. This new `scrollable` setting
helps us to prevent this behaviour.

A tweak has also been made into imgui, to allow window border to have
a different color than separators.