Adding validation to the add script button
Previously, it was possible to add an invalid script to an actor by
typing an incorrect script name in the script text field.

This has been fixed by adding a validation function (lambda) that is
updating the `Add Script` button state (Enabled or disabled) depending
on the content entered in the text field.