Improved actor creation menu
The actor creation menu had duplicated code. It actually was implemented
twice (In `MenuBar` and `Hierarchy`). To avoid that, an `OvEditor/Utils/`
folder has been added with an `ActorCreationMenu` class. This class
exposes a function that generates an actor creation menu.

It is now also possible to create a new actor by right-clicking onto the
`Root` node in the hierarchy.