Use implicitely our math in ImGui
This will allow use to forget about the `Converter` class on OvUI and use `ImVec2` and `ImVec4` seemlessly with our `OvMaths::FVector2` and `OvMaths::FVector4`.

The only tradeoff is the Color struct. I had to make some overloads like `operator==(const OvMaths::FVector4)` and `operator=(const OvMahts::FVector4)` to allow the affectation of a color from a vector4. We could instead use a specialized method in Color struct like `SetColor(const OvMaths::FVector4)` for example if we think it would be more appropriate and less error-prone when using Color and vector4.

Other than that, using our custom maths inside ImGui methods are seemless and works as intended.